V8 fixed vector_sink_int64
fixed problem with swig and return the long long int type vector in 
data() function. copied templates and result from ecss

+++ include/flaress/vector_sink_int64.h
@@ -52,7 +52,7 @@ namespace gr {
 
       //! Clear the data and tags containers.
       virtual void reset() = 0;
-      virtual std::vector<long long int> data() const = 0;
+      virtual std::vector<long int> data() const = 0;
       virtual std::vector<tag_t> tags() const = 0;
     };
